made flyout reusuable - although need to change behaviour so they can be toggled between each one from top nav
