+++ www/data/BlogPosts.json
@@ -1,0 +1,50 @@
+[
+  {
+    "type": "Blog post",
+    "title": "A web crawler handling millions of API requests.",
+    "description": "See how Monitoro built an automated a scraping platform using Supabase.",
+    "imgUrl": "images/case-study-monitoro.jpg",
+    "logoUrl": "images/logos/monitoro.png",
+    "organization": "Monitoro",
+    "url": "/casestudies/monitoro",
+    "postMeta": {
+      "name" : "Paul Copplestone",
+      "avatarUrl": "https://avatars0.githubusercontent.com/u/10214025?v=4",
+      "publishDate": "Mar 16, 2020",
+      "readLength": 6
+    },
+    "ctaText": "View case study"
+  },
+  {
+    "type": "Blog post",
+    "title": "A no-code website builder, deployed in seven days.",
+    "description": "See how Tayfa went from idea to paying customer in less than 30 days.",
+    "imgUrl": "images/case-study-tayfab.jpg",
+    "logoUrl": "images/logos/monitoro.png",
+    "organization": "Tayfab",
+    "url": "/casestudies/monitoro",
+    "postMeta": {
+      "name" : "Paul Copplestone",
+      "avatarUrl": "https://avatars0.githubusercontent.com/u/10214025?v=4",
+      "publishDate": "Mar 27, 2020",
+      "readLength": 5
+    },
+    "ctaText": "View case study"
+  },
+  {
+    "type": "Blog post",
+    "title": "Counter-fraud watchlists for the fintech industry.",
+    "description": "See how Xendit use Supabase to build a full-text search engine.",
+    "imgUrl": "images/case-study-xendit.jpg",
+    "logoUrl": "images/logos/xendit.png",
+    "organization": "Llama Lab",
+    "url": "/casestudies/monitoro",
+    "postMeta": {
+      "name" : "Paul Copplestone",
+      "avatarUrl": "https://avatars0.githubusercontent.com/u/10214025?v=4",
+      "publishDate": "Mar 20, 2020",
+      "readLength": 10
+    },
+    "ctaText": "View case study"
+  }
+]
